refactor(api): auth and user
