+++ apps/api/src/auth/auth.constants.ts
@@ -4,12 +4,3 @@ export const REFRESH_TOKEN_SECRET = "REFRESH_TOKEN_SECRET";
 export const JWT_KEY = "jwt";
 export const REFRESH_TOKEN_KEY = "refreshToken";
 export const IS_PUBLIC_KEY = "isPublic";
-
-export const MESSAGES = {
-  alreadyExist: {
-    email: "Looks like this account already exist, please check your mailbox.",
-  },
-  incorrect: {
-    password: "Incorrect Password. YOU SHALL NOT PASS!",
-  },
-};
