+++ apps/api/src/core/constants/message.constants.ts
@@ -1,0 +1,11 @@
+export const MESSAGES = {
+  notFound: {
+    user: "Oops! user not found.",
+  },
+  alreadyExist: {
+    email: "Looks like this account already exist, please check your mailbox.",
+  },
+  incorrect: {
+    password: "Incorrect Password. YOU SHALL NOT PASS!",
+  },
+};
