+++ apps/api/src/core/interceptors/index.ts
@@ -1,1 +1,1 @@
-export * from "./Serialize.interceptor";
+export * from "./serialize.interceptor";
