+++ apps/api/src/users/dto/update-user.dto.ts
@@ -1,5 +1,5 @@
-import { PartialType } from "@nestjs/swagger";
+import { OmitType } from "@nestjs/swagger";
 
 import { CreateUserDto } from "./create-user.dto";
 
-export class UpdateUserDto extends PartialType(CreateUserDto) {}
+export class UpdateUserDto extends OmitType(CreateUserDto, ["email"]) {}
