--- apps/api/src/users/users.constants.ts
@@ -1,5 +1,0 @@
-export const MESSAGES = {
-  notFound: {
-    user: "Oops! user not found.",
-  },
-};
