+++ apps/api/src/users/users.service.ts
@@ -2,10 +2,10 @@ import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model } from "mongoose";
 
+import { MESSAGES } from "../core/constants";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import { User } from "./schemas/user.schema";
-import { MESSAGES } from "./users.constants";
 
 @Injectable()
 export class UsersService {
@@ -27,10 +27,14 @@ export class UsersService {
   }
 
   async findOneByEmail(email: string) {
-    const user = await this.userModel.findOne({ email }).exec();
+    const user = await this.queryOneByEmail(email);
     if (!user) throw new NotFoundException(MESSAGES.notFound.user);
 
     return user;
+  }
+
+  queryOneByEmail(email: string) {
+    return this.userModel.findOne({ email }).exec();
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
